extract ValueBar to packages/common
