+++ client/packages/common/src/ui/components/charts/ValueBar/index.ts
@@ -1,0 +1,1 @@
+export * from './ValueBar';
