+++ client/packages/common/src/ui/components/charts/index.ts
@@ -29,3 +29,5 @@ export {
   XAxis,
   YAxis,
 };
+
+export * from './ValueBar';
