+++ client/packages/requisitions/src/RequestRequisition/DetailView/RequestLineEdit/ItemCharts/StockDistribution.tsx
@@ -4,6 +4,7 @@ import {
   Box,
   Tooltip,
   Typography,
+  ValueBar,
 } from '@openmsupply-client/common';
 import { useFormatNumber, useTranslation } from '@common/intl';
 import { useRequest } from '../../../api';
@@ -14,60 +15,11 @@ export interface StockDistributionProps {
   suggestedQuantity?: number;
 }
 
-const MIN_FLEX_BASIS_TO_SHOW_LABEL = 10;
-const MIN_FLEX_BASIS_TO_SHOW_VALUE = 5;
 const MIN_MC_WIDTH_TO_SHOW_TEXT = 5;
 
 const Divider = () => (
   <Box sx={{ backgroundColor: 'gray.dark', width: '1px', height: '45px' }} />
 );
-
-const ValueBar = ({
-  value,
-  total,
-  label,
-  colour,
-}: {
-  value: number;
-  total: number;
-  label: string;
-  colour: string;
-}) => {
-  const formatNumber = useFormatNumber();
-  if (value === 0) return null;
-
-  const flexBasis = Math.min(Math.round((100 * value) / total), 100);
-
-  return (
-    <>
-      <Tooltip title={`${label}: ${formatNumber.round(value)}`} placement="top">
-        <Box flexBasis={`${flexBasis}%`} flexGrow={1}>
-          <Box sx={{ backgroundColor: colour, height: '20px' }} />
-          <Box
-            style={{
-              textAlign: 'end',
-              paddingRight: 10,
-              paddingTop: 10,
-            }}
-          >
-            {flexBasis > MIN_FLEX_BASIS_TO_SHOW_LABEL ? (
-              <Typography
-                fontSize={12}
-                style={{ textOverflow: 'ellipsis', height: 20 }}
-              >
-                {label}
-              </Typography>
-            ) : null}
-            {flexBasis > MIN_FLEX_BASIS_TO_SHOW_VALUE ? (
-              <Typography fontSize={12}>{formatNumber.round(value)}</Typography>
-            ) : null}
-          </Box>
-        </Box>
-      </Tooltip>
-      <Divider />
-    </>
-  );
-};
 
 const MonthlyConsumption = ({
   month,
