+++ client/packages/requisitions/src/ResponseRequisition/DetailView/ReponseStats/ResponseStoreStats.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { useFormatNumber, useTranslation } from '@common/intl';
-import { Box, Tooltip, Typography } from '@openmsupply-client/common';
+import { useTranslation } from '@common/intl';
+import { Box, ValueBar } from '@openmsupply-client/common';
 
 export interface ResponseStoreStatsProps {
   stockOnHand: number;
@@ -10,59 +10,9 @@ export interface ResponseStoreStatsProps {
   otherRequestedQuantity: number;
 }
 
-const MIN_FLEX_BASIS_TO_SHOW_LABEL = 10;
-const MIN_FLEX_BASIS_TO_SHOW_VALUE = 5;
-
 const Divider = () => (
   <Box sx={{ backgroundColor: 'gray.dark', width: '1px', height: '45px' }} />
 );
-
-const ValueBar = ({
-  value,
-  total,
-  label,
-  colour,
-}: {
-  value: number;
-  total: number;
-  label: string;
-  colour: string;
-}) => {
-  const formatNumber = useFormatNumber();
-  if (value === 0) return null;
-
-  const flexBasis = Math.min(Math.round((100 * value) / total), 100);
-
-  return (
-    <>
-      <Tooltip title={`${label}: ${formatNumber.round(value)}`} placement="top">
-        <Box flexBasis={`${flexBasis}%`} flexGrow={1}>
-          <Box sx={{ backgroundColor: colour, height: '20px' }} />
-          <Box
-            style={{
-              textAlign: 'end',
-              paddingRight: 10,
-              paddingTop: 10,
-            }}
-          >
-            {flexBasis > MIN_FLEX_BASIS_TO_SHOW_LABEL ? (
-              <Typography
-                fontSize={12}
-                style={{ textOverflow: 'ellipsis', height: 20 }}
-              >
-                {label}
-              </Typography>
-            ) : null}
-            {flexBasis > MIN_FLEX_BASIS_TO_SHOW_VALUE ? (
-              <Typography fontSize={12}>{formatNumber.round(value)}</Typography>
-            ) : null}
-          </Box>
-        </Box>
-      </Tooltip>
-      <Divider />
-    </>
-  );
-};
 
 export const ResponseStoreStats: React.FC<ResponseStoreStatsProps> = ({
   stockOnHand,
@@ -77,7 +27,9 @@ export const ResponseStoreStats: React.FC<ResponseStoreStatsProps> = ({
 
   const predictedStockPercent =
     predictedStockLevels < totalRequested
-      ? `${Math.round((100 * predictedStockLevels) / totalRequested).toString()}%`
+      ? `${Math.round(
+          (100 * predictedStockLevels) / totalRequested
+        ).toString()}%`
       : '100%';
   const requestedPercent =
     totalRequested < predictedStockLevels
@@ -133,11 +85,7 @@ export const ResponseStoreStats: React.FC<ResponseStoreStatsProps> = ({
         }}
       >
         <>
-          <Box
-            display="flex"
-            alignItems="flex-start"
-            width={requestedPercent}
-          >
+          <Box display="flex" alignItems="flex-start" width={requestedPercent}>
             <Divider />
             <ValueBar
               value={requestedQuantity}
